LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/revert-val-007-expected.txt


FAIL A @keyframe animation with revert works when applied to multiple identical elements assert_equals: expected "21.440000534057617px" but got "0px"
